Removed Postgres specific from doobie-core tests
